+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 * Added `Lightness` property to the `ColorWheel` control, used
   to set the lightness of the generated color and to avoid
   resetting of lightness when using the `ColorEditorManager`
+* Added `Alpha` property to the `ColorWheel` control, used to
+  set the alpha channel of the generated color and to avoid
+  resetting of alpha when using the `ColorEditorManager`
 * Added `DisplayLightness` property to the `ColorWheel` control.
   When set, the wheel is drawn using the specified lightness,
   otherwise it draws the wheel at 50% (default)
@@ -39,7 +42,7 @@
 * Changing the `LightnessColorSlider.Value` property no longer
   updates the `Color` property
 * The `ColorManagerEditor` does a better job of syncing values
-  between attached controls   
+  between attached controls
 
 ### 1.8.0